Follow links to other markdown files
Add a `local-file-requested` signal to the preview webview. This
signal will be emited when the user clicks on a link starting with
`file://`.

The preview handler connects a callback to that signal, and
then calls `load_file(uri)` on the main window, which opens the linked
file (if Apostrophe can open it).

Solves issue #269